get rid of c_screen_data for good
A 1000x1000 array 'just in case' looked like a weird and wasteful
approach to storage to begin with; but forgetting to stop allocating
so many megabytes of memory _after it's no longer in use_?  Gimme a
break!  It was misleading to leave it there, too.  I actually assumed
it was in use at first.